Ajuste do Mangá Service e Dao.
Mudança do tipo de retorno da função update.

+++ api/ohara/src/main/java/br/com/ohara/dao/MangaDao.java
@@ -41,7 +41,7 @@ public interface MangaDao {
 			" quantidadeVolumes = :quantidadeVolumes" + 
 			" quantidadeCapitulos = :quantidadeCapitulos" +
 			" where idManga = :idManga;")
-	Long update(@BindBean Manga manga);
+	Integer update(@BindBean Manga manga);
 	
 	@SqlUpdate("delete " +
 			" from manga " +
